Divide Landpresenter into 3 parts.
